python server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
faster-whisper==1.0.3
openai-whisper==20231117
